Updating Spring Security and JWT Configuration

+++ src/main/java/com/example/Book_Store/repository/CustomerLoginRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CustomerLoginRepository extends JpaRepository<CustomerLogin, Long> {
+public interface CustomerLoginRepository extends JpaRepository<CustomerLogin, Integer> {
     boolean existsByEmail(String email);
 
     boolean existsByPassword(String password);
